# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotel Metasearch API",
    "description": "Compare hotel prices across 21 providers in real-time. Searches Booking.com, Agoda, Hotels.com, Expedia, Trip.com, Priceline, Kayak, Trivago, and 13 more providers simultaneously.",
    "version": "2.0.0"
  },
  "paths": {
    "/v2/hotels/search": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Search hotels across all providers",
        "operationId": "search_hotels_v2_hotels_search_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "description": "City name",
              "title": "City"
            },
            "description": "City name"
          },
          {
            "name": "check_in",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Check-in date YYYY-MM-DD",
              "title": "Check In"
            },
            "description": "Check-in date YYYY-MM-DD"
          },
          {
            "name": "check_out",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Check-out date YYYY-MM-DD",
              "title": "Check Out"
            },
            "description": "Check-out date YYYY-MM-DD"
          },
          {
            "name": "adults",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9,
              "minimum": 1,
              "default": 2,
              "title": "Adults"
            }
          },
          {
            "name": "rooms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "default": 1,
              "title": "Rooms"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "USD",
              "title": "Currency"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated provider names",
              "title": "Sources"
            },
            "description": "Comma-separated provider names"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Price"
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price"
            }
          },
          {
            "name": "min_stars",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Stars"
            }
          },
          {
            "name": "max_stars",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Stars"
            }
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 10,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rating"
            }
          },
          {
            "name": "free_cancel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Free Cancel"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort: price, rating, stars, reviews",
              "default": "price",
              "title": "Sort By"
            },
            "description": "Sort: price, rating, stars, reviews"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/hotels/cheapest": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Find the single cheapest hotel",
        "operationId": "cheapest_hotel_v2_hotels_cheapest_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          },
          {
            "name": "check_in",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check In"
            }
          },
          {
            "name": "check_out",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Out"
            }
          },
          {
            "name": "adults",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9,
              "minimum": 1,
              "default": 2,
              "title": "Adults"
            }
          },
          {
            "name": "rooms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "default": 1,
              "title": "Rooms"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "USD",
              "title": "Currency"
            }
          },
          {
            "name": "min_stars",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Stars"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sources"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/hotels/compare": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Compare one hotel across all providers",
        "operationId": "compare_hotel_v2_hotels_compare_get",
        "parameters": [
          {
            "name": "hotel_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hotel name to compare",
              "title": "Hotel Name"
            },
            "description": "Hotel name to compare"
          },
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          },
          {
            "name": "check_in",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check In"
            }
          },
          {
            "name": "check_out",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Out"
            }
          },
          {
            "name": "adults",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9,
              "minimum": 1,
              "default": 2,
              "title": "Adults"
            }
          },
          {
            "name": "rooms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "default": 1,
              "title": "Rooms"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "USD",
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/hotels/deals": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Find price deals across providers",
        "operationId": "hotel_deals_v2_hotels_deals_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          },
          {
            "name": "check_in",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check In"
            }
          },
          {
            "name": "check_out",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Out"
            }
          },
          {
            "name": "adults",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9,
              "minimum": 1,
              "default": 2,
              "title": "Adults"
            }
          },
          {
            "name": "rooms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "default": 1,
              "title": "Rooms"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "USD",
              "title": "Currency"
            }
          },
          {
            "name": "min_savings",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 100,
              "minimum": 0,
              "description": "Min savings % to include",
              "default": 0,
              "title": "Min Savings"
            },
            "description": "Min savings % to include"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}